LARS/Tripletex/requirements.txt
fastapi
uvicorn[standard]
google-adk
google-genai
pdfplumber
Pillow
requests
python-dotenv
